+++ lib/insta_assets_picker.dart
@@ -2,7 +2,11 @@ library insta_assets_picker;
 
 export 'src/assets_picker.dart';
 export 'src/insta_assets_crop_controller.dart'
-    show InstaAssetsExportDetails, InstaAssetsCrop;
+    show InstaAssetsExportDetails, InstaAssetsCropData;
 export 'package:wechat_assets_picker/wechat_assets_picker.dart'
-    show AssetEntity, AssetEntityImageProvider, DefaultAssetPickerProvider;
+    show
+        AssetEntity,
+        AssetEntityImageProvider,
+        DefaultAssetPickerProvider,
+        SpecialItemPosition;
 export 'package:wechat_assets_picker/src/delegates/asset_picker_text_delegate.dart';