+++ lib/src/assets_picker.dart
@@ -7,6 +7,31 @@ import 'package:wechat_assets_picker/wechat_assets_picker.dart';
 
 const _kGridCount = 4;
 const _kInitializeDelayDuration = Duration(milliseconds: 250);
+const kDefaultInstaCropRatios = [1.0, 4 / 5];
+
+class InstaAssetCropDelegate {
+  const InstaAssetCropDelegate({
+    this.preferredSize = 1080,
+    this.cropRatios = kDefaultInstaCropRatios,
+  });
+
+  /// The param [preferredSize] is used to produce higher quality cropped image.
+  /// Keep in mind that the higher this value is, the heavier the cropped image will be.
+  ///
+  /// This value while be used as such
+  /// ```dart
+  /// preferredSize = (preferredSize / scale).round()
+  /// ```
+  ///
+  /// Defaults to `1080`, like instagram.
+  final double preferredSize;
+
+  /// The param [cropRatios] provided the list of crop ratios that can be set
+  /// from the crop view.
+  ///
+  /// Defaults to `[1/1, 4/5]` like instagram.
+  final List<double> cropRatios;
+}
 
 class InstaAssetPicker {
   InstaAssetPickerBuilder? builder;
@@ -61,6 +86,9 @@ class InstaAssetPicker {
   /// Set [onPermissionDenied] to manually handle the denied permission error.
   /// The default behavior is to open a [ScaffoldMessenger].
   ///
+  /// Crop options
+  /// - Set [cropDelegate] to customize the display and export of crops.
+  ///
   /// Those arguments are used by [InstaAssetPickerBuilder]
   ///
   /// - Set [provider] of type [DefaultAssetPickerProvider] to specifies picker options.
@@ -80,10 +108,6 @@ class InstaAssetPicker {
   /// - Set [closeOnComplete] to specifies if the picker should be closed
   /// after assets selection confirmation.
   ///
-  /// - Set [isSquareDefaultCrop] to specifies if the crop view should be initialized
-  /// on square mode.
-  /// Defaults to `false`, like instagram.
-  ///
   /// - The [onCompleted] callback is called when the assets selection is confirmed.
   /// It will as argument a [Stream] with exportation details [InstaAssetsExportDetails].
   ///
@@ -97,6 +121,9 @@ class InstaAssetPicker {
     Function(BuildContext context, String delegateDescription)?
         onPermissionDenied,
 
+    /// Crop options
+    InstaAssetCropDelegate cropDelegate = const InstaAssetCropDelegate(),
+
     /// InstaAssetPickerBuilder options
     int gridCount = _kGridCount,
     required DefaultAssetPickerProvider provider,
@@ -104,10 +131,11 @@ class InstaAssetPicker {
     AssetPickerTextDelegate? textDelegate,
     String? title,
     bool closeOnComplete = false,
-    bool isSquareDefaultCrop = false,
     required Function(Stream<InstaAssetsExportDetails> exportDetails)
         onCompleted,
     Widget Function(BuildContext, bool)? loadingIndicatorBuilder,
+    Widget? Function(BuildContext, AssetPathEntity?, int)? specialItemBuilder,
+    SpecialItemPosition? specialItemPosition,
   }) async {
     assert(provider.requestType == RequestType.image,
         'Only images can be shown in the picker for now');
@@ -135,8 +163,10 @@ class InstaAssetPicker {
       textDelegate: text,
       loadingIndicatorBuilder: loadingIndicatorBuilder,
       closeOnComplete: closeOnComplete,
-      isSquareDefaultCrop: isSquareDefaultCrop,
+      cropDelegate: cropDelegate,
       onCompleted: onCompleted,
+      specialItemBuilder: specialItemBuilder,
+      specialItemPosition: specialItemPosition,
     );
 
     return AssetPicker.pickAssetsWithDelegate(
@@ -158,6 +188,9 @@ class InstaAssetPicker {
   /// Set [onPermissionDenied] to manually handle the denied permission error.
   /// The default behavior is to open a [ScaffoldMessenger].
   ///
+  /// Crop options
+  /// - Set [cropDelegate] to customize the display and export of crops.
+  ///
   /// Those arguments are used by [InstaAssetPickerBuilder]
   ///
   /// - Set [gridCount] to specifies the number of assets in the cross axis.
@@ -173,10 +206,6 @@ class InstaAssetPicker {
   ///
   /// - Set [closeOnComplete] to specifies if the picker should be closed
   /// after assets selection confirmation.
-  ///
-  /// - Set [isSquareDefaultCrop] to specifies if the crop view should be initialized
-  /// on square mode.
-  /// Defaults to `false`, like instagram.
   ///
   /// - The [onCompleted] callback is called when the assets selection is confirmed.
   /// It will as argument a [Stream] with exportation details [InstaAssetsExportDetails].
@@ -217,13 +246,15 @@ class InstaAssetPicker {
     Function(BuildContext context, String delegateDescription)?
         onPermissionDenied,
 
+    /// Crop options
+    InstaAssetCropDelegate cropDelegate = const InstaAssetCropDelegate(),
+
     /// InstaAssetPickerBuilder options
     int gridCount = _kGridCount,
     ThemeData? pickerTheme,
     AssetPickerTextDelegate? textDelegate,
     String? title,
     bool closeOnComplete = false,
-    bool isSquareDefaultCrop = false,
     required Function(Stream<InstaAssetsExportDetails> exportDetails)
         onCompleted,
     Widget Function(BuildContext, bool)? loadingIndicatorBuilder,
@@ -238,6 +269,8 @@ class InstaAssetPicker {
     bool sortPathsByModifiedDate = false,
     FilterOptionGroup? filterOptions,
     Duration initializeDelayDuration = _kInitializeDelayDuration,
+    Widget? Function(BuildContext, AssetPathEntity?, int)? specialItemBuilder,
+    SpecialItemPosition? specialItemPosition,
   }) async {
     final locale = Localizations.maybeLocaleOf(context);
     final text = textDelegate ?? assetPickerTextDelegateFromLocale(locale);
@@ -274,8 +307,10 @@ class InstaAssetPicker {
       textDelegate: text,
       loadingIndicatorBuilder: loadingIndicatorBuilder,
       closeOnComplete: closeOnComplete,
-      isSquareDefaultCrop: isSquareDefaultCrop,
+      cropDelegate: cropDelegate,
       onCompleted: onCompleted,
+      specialItemBuilder: specialItemBuilder,
+      specialItemPosition: specialItemPosition,
     );
 
     return AssetPicker.pickAssetsWithDelegate(