+++ lib/src/insta_assets_crop_controller.dart
@@ -1,9 +1,8 @@
 import 'dart:io';
 
 import 'package:flutter/material.dart';
-import 'package:image_crop/image_crop.dart' hide CropState;
-import 'package:insta_assets_picker/src/custom_packages/image_crop/crop.dart'
-    show CropState, CropInternal;
+import 'package:fraction/fraction.dart';
+import 'package:insta_assets_crop/insta_assets_crop.dart';
 import 'package:insta_assets_picker/insta_assets_picker.dart';
 import 'package:wechat_assets_picker/wechat_assets_picker.dart';
 
@@ -13,7 +12,7 @@ import 'package:wechat_assets_picker/wechat_assets_picker.dart';
 class InstaAssetsCropSingleton {
   const InstaAssetsCropSingleton._();
 
-  static List<InstaAssetsCrop> cropParameters = [];
+  static List<InstaAssetsCropData> cropParameters = [];
 }
 
 /// Contains all the parameters of the exportation
@@ -39,7 +38,7 @@ class InstaAssetsExportDetails {
 }
 
 /// The crop parameters state, can be used at exportation or to load the crop view
-class InstaAssetsCrop {
+class InstaAssetsCropData {
   final AssetEntity asset;
   final CropInternal? cropParam;
 
@@ -47,18 +46,18 @@ class InstaAssetsCrop {
   final double scale;
   final Rect? area;
 
-  const InstaAssetsCrop({
+  const InstaAssetsCropData({
     required this.asset,
     required this.cropParam,
     this.scale = 1.0,
     this.area,
   });
 
-  static InstaAssetsCrop fromState({
+  static InstaAssetsCropData fromState({
     required AssetEntity asset,
     required CropState? cropState,
   }) {
-    return InstaAssetsCrop(
+    return InstaAssetsCropData(
       asset: asset,
       cropParam: cropState?.internalParameters,
       scale: cropState?.scale ?? 1.0,
@@ -69,11 +68,11 @@ class InstaAssetsCrop {
 
 /// The controller that handles the exportation and save the state of the selected assets crop parameters
 class InstaAssetsCropController {
-  InstaAssetsCropController(this.keepMemory, bool isSquareDefaultCrop)
-      : isSquare = ValueNotifier<bool>(isSquareDefaultCrop);
+  InstaAssetsCropController(this.keepMemory, this.cropDelegate)
+      : cropRatioIndex = ValueNotifier<int>(0);
 
-  /// Whether the crop view aspect ratio is 1 or 4/5
-  late final ValueNotifier<bool> isSquare;
+  /// The index of the selected aspectRatio among the possibilities
+  final ValueNotifier<int> cropRatioIndex;
 
   /// Whether the image in the crop view is loaded
   final ValueNotifier<bool> isCropViewReady = ValueNotifier<bool>(false);
@@ -82,29 +81,51 @@ class InstaAssetsCropController {
   final ValueNotifier<AssetEntity?> previewAsset =
       ValueNotifier<AssetEntity?>(null);
 
+  /// Options related to crop
+  final InstaAssetCropDelegate cropDelegate;
+
   /// List of all the crop parameters set by the user
-  List<InstaAssetsCrop> _cropParameters = [];
+  List<InstaAssetsCropData> _cropParameters = [];
 
   /// Whether if [_cropParameters] should be saved in the cache to use when the picker
   /// is open with [InstaAssetPicker.restorableAssetsPicker]
-  bool keepMemory = false;
+  final bool keepMemory;
 
   dispose() {
     clear();
     isCropViewReady.dispose();
-    isSquare.dispose();
+    cropRatioIndex.dispose();
     previewAsset.dispose();
   }
 
-  double get aspectRatio => isSquare.value ? 1 : 4 / 5;
+  double get aspectRatio {
+    assert(cropDelegate.cropRatios.isNotEmpty,
+        'The list of supported crop ratios cannot be empty.');
+    return cropDelegate.cropRatios[cropRatioIndex.value];
+  }
+
+  String get aspectRatioString {
+    final r = aspectRatio;
+    if (r == 1) return '1:1';
+    return Fraction.fromDouble(r).reduce().toString().replaceFirst('/', ':');
+  }
+
+  /// Set the next available index as the selected crop ratio
+  void nextCropRatio() {
+    if (cropRatioIndex.value < cropDelegate.cropRatios.length - 1) {
+      cropRatioIndex.value = cropRatioIndex.value + 1;
+    } else {
+      cropRatioIndex.value = 0;
+    }
+  }
 
   /// Use [_cropParameters] when [keepMemory] is `false`, otherwise use [InstaAssetsCropSingleton.cropParameters]
-  List<InstaAssetsCrop> get cropParameters =>
+  List<InstaAssetsCropData> get cropParameters =>
       keepMemory ? InstaAssetsCropSingleton.cropParameters : _cropParameters;
 
   /// Save the list of crop parameters
   /// if [keepMemory] save list memory or simply in the controller
-  void updateStoreCropParam(List<InstaAssetsCrop> list) {
+  void updateStoreCropParam(List<InstaAssetsCropData> list) {
     if (keepMemory) {
       InstaAssetsCropSingleton.cropParameters = list;
     } else {
@@ -124,7 +145,7 @@ class InstaAssetsCropController {
     CropState? saveCropState,
     List<AssetEntity> selectedAssets,
   ) {
-    final List<InstaAssetsCrop> newList = [];
+    final List<InstaAssetsCropData> newList = [];
 
     for (final asset in selectedAssets) {
       // get the already saved crop parameters if exists
@@ -133,14 +154,15 @@ class InstaAssetsCropController {
       // if it is the asseet to save & the crop parameters exists
       if (asset == saveAsset && saveAsset != null) {
         // add the new parameters
-        newList.add(InstaAssetsCrop.fromState(
+        newList.add(InstaAssetsCropData.fromState(
           asset: saveAsset,
           cropState: saveCropState,
         ));
         // if it is not the asset to save and no crop parameter exists
       } else if (savedCropAsset == null) {
         // set empty crop parameters
-        newList.add(InstaAssetsCrop.fromState(asset: asset, cropState: null));
+        newList
+            .add(InstaAssetsCropData.fromState(asset: asset, cropState: null));
       } else {
         // keep existing crop parameters
         newList.add(savedCropAsset);
@@ -151,8 +173,8 @@ class InstaAssetsCropController {
     updateStoreCropParam(newList);
   }
 
-  /// Returns the crop parametes [InstaAssetsCrop] of the given asset
-  InstaAssetsCrop? get(AssetEntity asset) {
+  /// Returns the crop parametes [InstaAssetsCropData] of the given asset
+  InstaAssetsCropData? get(AssetEntity asset) {
     if (cropParameters.isEmpty) return null;
     final index = cropParameters.indexWhere((e) => e.asset == asset);
     if (index == -1) return null;
@@ -190,27 +212,27 @@ class InstaAssetsCropController {
       if (file == null) {
         throw 'error file is null';
       }
-
+      
       if (asset.type == AssetType.video) {
         croppedFiles.add(file);
       } else {
-        // makes the sample file to not be too small
-        final sampledFile = await ImageCrop.sampleImage(
-          file: file,
-          preferredSize: (1024 / scale).round(),
-        );
+                // makes the sample file to not be too small
+      final sampledFile = await InstaAssetsCrop.sampleImage(
+        file: file,
+        preferredSize: (cropDelegate.preferredSize / scale).round(),
+      );
 
-        if (area == null) {
-          croppedFiles.add(sampledFile);
-        } else {
-          // crop the file with the area selected
-          final croppedFile =
-              await ImageCrop.cropImage(file: sampledFile, area: area);
-          // delete the not needed sample file
-          sampledFile.delete();
+      if (area == null) {
+        croppedFiles.add(sampledFile);
+      } else {
+        // crop the file with the area selected
+        final croppedFile =
+            await InstaAssetsCrop.cropImage(file: sampledFile, area: area);
+        // delete the not needed sample file
+        sampledFile.delete();
 
-          croppedFiles.add(croppedFile);
-        }
+        croppedFiles.add(croppedFile);
+      }
       }
 
       // increase progress